Práticas_de_Programação/Concessionaria_3.0/src/br/com/pratica/projeto/model/Fabricante.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package br.com.pratica.projeto.model;

/**
 *
 * @author Nellson
 */
public class Fabricante {
    
    private long id;
    private String nome;
    private int cnpj;

    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }

    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    public int getCnpj() {
        return cnpj;
    }

    public void setCnpj(int cnpj) {
        this.cnpj = cnpj;
    }    

    @Override
    public String toString() {
        return nome + " - CNPJ: " + cnpj;
    }
}
